fixed problem with manhattan
Similar to the previous commit but this time for manhattan.
previously we only calculated the mahattan weight as the sum of moves for each tile to get to its solution spot.
Now we take the depth aka the number of moves into account and add it to the weight but to manhattan. Hence f(n) = g(n) + h(n)